requirements-dev.txt
black>=21.7b0
eth-brownie>=1.16.2,<2.0.0
